Add tests for express server using jest
